==> 16_Anova_test_WL_stdev_landtype.ipynb <==
### Simple one-way ANOVA test to analyse the relationship between the water level variability and peat type

**Motivation for analysis**
- Visually, there seems to be a strong correspondence between water level variation and peat vegetation type, with higher variation over palm swamp regions. 

**Inputs**
- land type map 
- standard deviation of the final derived water level timeseries

**Method**
- Calculate water level standard deviation
- Calculate one-way ANOVA (f-test) between the peat swamp type (hardwood/palm) and the water level variation



from scipy import stats
from scipy.stats import f_oneway
from statsmodels.formula.api import ols
from statsmodels.stats.anova import anova_lm
import pandas as pd
import xarray as xr
import numpy as np
xr.set_options(cmap_sequential='viridis')

ALOS_OUT = 'filepath'

# Landtype map
lt_map = xr.open_dataset(ALOS_OUT + 'landtype_100m.nc')
lt_map = lt_map.where(lt_map.isin([4,5]))
# water level daily timeseries
wl_final = xr.open_dataset(ALOS_OUT + 'WL_daily_final.nc')

# mean, min, max and stdev
WL_mean = xr.open_dataset(ALOS_OUT + 'WL_mean.nc')
WL_min = xr.open_dataset(ALOS_OUT + 'WL_min.nc')
WL_max = xr.open_dataset(ALOS_OUT + 'WL_max.nc')
WL_stdev = xr.open_dataset(ALOS_OUT + 'WL_stdev.nc')

# One way anova between the peat swamp type and the water level standard deviation
def calculate_anova(lt_map,wl_stat, stat_name):
    d1 = lt_map['type'].values
    d2 = wl_stat['water_level'].values

    # removing nan values
    d1 = d1[~np.isnan(d1)]
    d2 = d2[~np.isnan(d2)]
    f_stat, pvalue = f_oneway(d1,d2)

    print ('The one-way ANOVA outputs for ' + stat_name + ' are: ')
    print ('f stat: ', str(f_stat))
    print ('p value: ', str(pvalue))
    
    return (d2)



d2 = calculate_anova(lt_map, WL_stdev, 'the standard deviation of the water level')
print (len(d2))

d2

calculate_anova(lt_map, WL_mean, 'the mean water level')

calculate_anova(lt_map, WL_min, 'the min water level')

calculate_anova(lt_map, WL_max, 'the max water level')

#### testing the eta statistic 
- eta calculates the effect size of the one way anova test (or eta squared, like R squared) - it measures the proportion of the total variance explained by the independent variable
- eta squared is sensitive to sample size
- larger the sample size, the larger the effect size
- important to mention the effect size and sample size together

%%time

# merging the land type and stdev arrays
ds_merged = xr.merge([lt_map['type'],WL_stdev['water_level']])
df = ds_merged.to_dataframe()


aov_table

%%time

# merging the land type and stdev arrays
ds_merged = xr.merge([lt_map['type'],WL_stdev['water_level']])
df = ds_merged.to_dataframe()

model = ols('water_level ~ C(type)', data = df).fit()

# anova test
aov_table = anova_lm(model, typ=2)

# effect size
# sum of squares between the groups
SSbetween = aov_table.loc['C(type)','sum_sq']
df_between = aov_table.loc['C(type)','df']

# sum of squares within the groups
SSwithin = aov_table.loc['Residual','sum_sq']
df_within = aov_table.loc['Residual','df']

eta_squared = SSbetween / (SSbetween + SSwithin)
eta_squared2 = 3828724/(3828724+36854100)
print (aov_table)
print ('Eta squared: ' + str(eta_squared))
print ('Eta squared partial: ' + str(eta_squared2))

**About 9.4% of the variance in peat swamp type is due to the water level variance (pvalue<<0.05)**

%%time

# merging the land type and stdev arrays
ds_merged = xr.merge([lt_map['type'],WL_min['water_level']])
df = ds_merged.to_dataframe()

model = ols('water_level ~ C(type)', data = df).fit()

# anova test
aov_table = anova_lm(model, typ=2)

# effect size
# sum of squares between the groups
SSbetween = aov_table.loc['C(type)','sum_sq']
df_between = aov_table.loc['C(type)','df']

# sum of squares within the groups
SSwithin = aov_table.loc['Residual','sum_sq']
df_within = aov_table.loc['Residual','df']

eta_squared = SSbetween / (SSbetween + SSwithin)

print (aov_table)
print ('\nEta squared: ' + str(eta_squared))

**None of the variance in peat swamp type is related to the minimum water level**

%%time

# merging the land type and stdev arrays
ds_merged = xr.merge([lt_map['type'],WL_max['water_level']])
df = ds_merged.to_dataframe()

model = ols('water_level ~ C(type)', data = df).fit()

# anova test
aov_table = anova_lm(model, typ=2)

# effect size
# sum of squares between the groups
SSbetween = aov_table.loc['C(type)','sum_sq']
df_between = aov_table.loc['C(type)','df']

# sum of squares within the groups
SSwithin = aov_table.loc['Residual','sum_sq']
df_within = aov_table.loc['Residual','df']

eta_squared = SSbetween / (SSbetween + SSwithin)


print (aov_table)
print ('\nEta squared: ' + str(eta_squared))

**About 12.5% of the variance in peat swamp type is due to the max water level (pvalue<<0.05)**

%%time

# merging the land type and stdev arrays
ds_merged = xr.merge([lt_map['type'],WL_mean['water_level']])
df = ds_merged.to_dataframe()

model = ols('water_level ~ C(type)', data = df).fit()

# anova test
aov_table = anova_lm(model, typ=2)

# effect size
# sum of squares between the groups
SSbetween = aov_table.loc['C(type)','sum_sq']
df_between = aov_table.loc['C(type)','df']

# sum of squares within the groups
SSwithin = aov_table.loc['Residual','sum_sq']
df_within = aov_table.loc['Residual','df']

eta_squared = SSbetween / (SSbetween + SSwithin)
print (aov_table)
print ('\nEta squared: ' + str(eta_squared))

**About 9.4% of the variance in peat swamp type is due to the mean water level (pvalue<<0.05)**

#### Mean value of max, mean and stdev for each peat swamp type

def calculate_mean_values(ds, variable_name):
    print ('Palm swamp ' + variable_name + str(np.round(ds.where(lt_map['type']==4)['water_level'].mean().values,2)) + ' cm')
    print ('Hardwood swamp ' + variable_name + str(np.round(ds.where(lt_map['type']==5)['water_level'].mean().values,2)) + ' cm')
                                                          

calculate_mean_values(WL_stdev,'standard deviation: ')
calculate_mean_values(WL_min,'min: ')
calculate_mean_values(WL_max,'max: ')
calculate_mean_values(WL_mean,'mean: ')

### The minimum value might not be of much significance here because the rainfall was significantly higher during this period than usual. It would be interesting to investigate if minimum water level becomes significant
